dev: support sequences of length 64
Former-commit-id: 323db250be03edb24feed449ff0d965704f2bf30

+++ src/spexs/trie/fullset.go
@@ -21,8 +21,8 @@ func (f *FullSet) Len() int {
 	return f.Count
 }
 
-func (f *FullSet) Iter() map[int]uint {
-	result := make(map[int]uint, len(f.Ref.Seqs))
+func (f *FullSet) Iter() map[int]uint64 {
+	result := make(map[int]uint64, len(f.Ref.Seqs))
 
 	for idx, pat := range f.Ref.Seqs {
 		result[idx] = 2<<byte(len(pat.Pat)) - 1
